Refactor `Books``
