+++ lib/cake/books.ex
@@ -3,123 +3,9 @@ defmodule Cake.Books do
   The Books context.
   """
 
-  import Ecto.Query, warn: false
   alias Cake.Books.Chunk
   alias Cake.Books.ParsedBook
   alias Cake.Repo
-
-  require Logger
-
-  @spec persist_books_and_chunks({ParsedBook.t(), [Chunk.t()]} | {term(), term()}) ::
-          {:ok, {ParsedBook.t(), [Chunk.t()]}} | {:error, any()}
-  def persist_books_and_chunks({%ParsedBook{file_hash: hash} = book, chunks})
-      when is_list(chunks) do
-    case Repo.one(from b in ParsedBook, where: b.file_hash == ^hash) do
-      %ParsedBook{} = existing ->
-        existing_chunks = Repo.all(from c in Chunk, where: c.parsed_book_id == ^existing.id)
-        Logger.debug("Skipping already-persisted book #{existing.title} (#{hash})")
-        {:ok, {existing, existing_chunks}}
-
-      nil ->
-        persist_books_and_chunks(book, chunks)
-    end
-  end
-
-  def persist_books_and_chunks({book, chunks}) do
-    {:error, {:invalid_input, %{book: book, chunks: chunks}}}
-  end
-
-  @spec persist_books_and_chunks(ParsedBook.t(), [Chunk.t()]) ::
-          {:ok, {ParsedBook.t(), [Chunk.t()]}} | {:error, any()}
-  def persist_books_and_chunks(%ParsedBook{} = book, chunks) when is_list(chunks) do
-    book
-    |> build_multi(chunks)
-    |> run_transaction(book)
-  end
-
-  defp build_multi(book, chunks) do
-    book_attrs =
-      book
-      |> Map.from_struct()
-      |> Map.drop([:__meta__, :id, :inserted_at, :updated_at, :chunks])
-
-    Ecto.Multi.new()
-    |> Ecto.Multi.insert(:book, ParsedBook.changeset(%ParsedBook{}, book_attrs))
-    |> Ecto.Multi.run(:chunks, fn repo, %{book: persisted_book} ->
-      insert_chunks(repo, persisted_book, chunks)
-    end)
-  end
-
-  defp insert_chunks(repo, persisted_book, chunks) do
-    now = DateTime.truncate(DateTime.utc_now(), :second)
-
-    chunks
-    |> Enum.with_index()
-    |> Enum.reduce_while({:ok, []}, fn {chunk, idx}, {:ok, acc} ->
-      case build_chunk_row(chunk, persisted_book.id, idx, now) do
-        {:ok, row} -> {:cont, {:ok, [row | acc]}}
-        {:error, _} = error -> {:halt, error}
-      end
-    end)
-    |> then(fn
-      {:ok, reversed_rows} -> bulk_insert_chunks(repo, Enum.reverse(reversed_rows))
-      {:error, _} = error -> error
-    end)
-  end
-
-  defp build_chunk_row(chunk, book_id, idx, now) do
-    chunk_attrs =
-      chunk
-      |> Map.from_struct()
-      |> Map.drop([:__meta__, :id, :inserted_at, :updated_at, :parsed_book])
-      |> Map.put(:parsed_book_id, book_id)
-      |> Map.put_new(:chunk_index, idx)
-
-    cs = Chunk.changeset(%Chunk{}, chunk_attrs)
-
-    if cs.valid? do
-      row =
-        cs.changes
-        |> Map.put(:inserted_at, now)
-        |> Map.put(:updated_at, now)
-
-      {:ok, row}
-    else
-      {:error, {:invalid_chunk, cs.errors, chunk_attrs}}
-    end
-  end
-
-  defp bulk_insert_chunks(repo, chunk_rows) do
-    {count, returned_rows} =
-      repo.insert_all(Chunk, chunk_rows,
-        returning: [
-          :id,
-          :parsed_book_id,
-          :page_number,
-          :chunk_index,
-          :section_title,
-          :text,
-          :word_count,
-          :char_count
-        ]
-      )
-
-    if count == length(chunk_rows) do
-      {:ok, returned_rows}
-    else
-      {:error, {:chunk_insert_count_mismatch, count, length(chunk_rows)}}
-    end
-  end
-
-  defp run_transaction(multi, book) do
-    case Repo.transaction(multi) do
-      {:ok, %{book: persisted_book, chunks: persisted_chunks}} ->
-        {:ok, {persisted_book, persisted_chunks}}
-
-      {:error, _step, reason, _changes_so_far} ->
-        {:error, {book.source_file_path, reason}}
-    end
-  end
 
   @doc """
   Returns the list of parsed_books.
@@ -235,90 +121,6 @@ defmodule Cake.Books do
   @spec list_chunks() :: [Chunk.t()]
   def list_chunks do
     Repo.all(Chunk)
-  end
-
-  @doc """
-  Fetches Chunk records for a list of OpenSearch hits, with `parsed_book`
-  preloaded. Returns chunks in the same order as the hits.
-  """
-  @spec chunks_for_hits(%Snap.Hits{} | list()) :: [Chunk.t()]
-  def chunks_for_hits(hits) do
-    ids = Enum.map(hits, fn hit -> hit.source["id"] end)
-
-    chunks_by_id =
-      Map.new(
-        Repo.all(from c in Chunk, where: c.id in ^ids, preload: :parsed_book),
-        fn chunk -> {chunk.id, chunk} end
-      )
-
-    ids
-    |> Enum.map(&Map.get(chunks_by_id, &1))
-    |> Enum.reject(&is_nil/1)
-  end
-
-  @doc """
-  Expands a list of retrieved chunks by fetching neighboring chunks from Postgres.
-
-  Given chunks returned by `chunks_for_hits/1`, this function:
-  1. Groups them by `parsed_book_id`
-  2. For each book, computes the union of chunk index ranges [index - offset, index + offset]
-  3. Merges overlapping ranges to avoid redundant queries
-  4. Fetches all chunks in those ranges from Postgres
-  5. Returns a deduplicated, ordered list with `:parsed_book` preloaded
-
-  The offset controls how many chunks on each side of a hit are included.
-  An offset of 2 means each hit brings in up to 4 neighbors (2 before, 2 after),
-  though in practice overlapping hits within the same book will merge into
-  contiguous windows.
-
-  ## Examples
-
-      # Expand each hit by 2 chunks on each side
-      chunks = Books.chunks_for_hits(hits)
-      expanded = Books.expand_with_neighbors(chunks, 2)
-
-  """
-  @spec expand_with_neighbors([Chunk.t()], non_neg_integer()) :: [Chunk.t()]
-  def expand_with_neighbors(chunks, offset)
-      when is_list(chunks) and is_integer(offset) and offset >= 0 do
-    chunks
-    |> Enum.group_by(& &1.parsed_book_id)
-    |> Enum.flat_map(fn {book_id, book_chunks} ->
-      fetch_neighbor_ranges(book_id, book_chunks, offset)
-    end)
-    |> Enum.uniq_by(& &1.id)
-  end
-
-  defp fetch_neighbor_ranges(book_id, book_chunks, offset) do
-    book_chunks
-    |> Enum.map(fn c -> {max(c.chunk_index - offset, 0), c.chunk_index + offset} end)
-    |> Enum.sort()
-    |> merge_ranges()
-    |> Enum.flat_map(fn {low, high} ->
-      Chunk.base_query()
-      |> Chunk.by_book(book_id)
-      |> where([c], c.chunk_index >= ^low and c.chunk_index <= ^high)
-      |> order_by([c], asc: c.chunk_index)
-      |> Repo.all()
-      |> Repo.preload(:parsed_book)
-    end)
-  end
-
-  # Merges a sorted list of {low, high} integer ranges into non-overlapping ranges.
-  # Adjacent ranges (e.g., {1, 3} and {4, 6}) are also merged since chunk indices
-  # are integers and index 3 and index 4 are contiguous.
-  defp merge_ranges([]), do: []
-
-  defp merge_ranges([first | rest]) do
-    Enum.reverse(
-      Enum.reduce(rest, [first], fn {low, high}, [{acc_low, acc_high} | tail] ->
-        if low <= acc_high + 1 do
-          [{acc_low, max(acc_high, high)} | tail]
-        else
-          [{low, high}, {acc_low, acc_high} | tail]
-        end
-      end)
-    )
   end
 
   @doc """
